+++ src/main/gen/com/aesirinteractive/angelscript/psi/AngelscriptTypeRef.java
@@ -4,6 +4,7 @@ package com.aesirinteractive.angelscript.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiReference;
 
 public interface AngelscriptTypeRef extends PsiElement {
 
@@ -15,5 +16,7 @@ public interface AngelscriptTypeRef extends PsiElement {
 
   @Nullable
   PsiElement getIdentifier();
+
+  @Nullable PsiReference getReference();
 
 }