+++ src/main/gen/com/aesirinteractive/angelscript/psi/impl/AngelscriptTypeRefImpl.java
@@ -8,11 +8,12 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.aesirinteractive.angelscript.AngelscriptTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import com.aesirinteractive.angelscript.AngelscriptTypeRefMixin;
 import com.aesirinteractive.angelscript.psi.*;
 import com.aesirinteractive.angelscript.AngelscriptPsiImplUtil;
+import com.intellij.psi.PsiReference;
 
-public class AngelscriptTypeRefImpl extends ASTWrapperPsiElement implements AngelscriptTypeRef {
+public class AngelscriptTypeRefImpl extends AngelscriptTypeRefMixin implements AngelscriptTypeRef {
 
   public AngelscriptTypeRefImpl(@NotNull ASTNode node) {
     super(node);