+++ src/main/gen/com/aesirinteractive/angelscript/psi/impl/AngelscriptVariableAccessExprImpl.java
@@ -8,16 +8,17 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.aesirinteractive.angelscript.AngelscriptTypes.*;
+import com.aesirinteractive.angelscript.AngelscriptVariableAccessExprMixin;
 import com.aesirinteractive.angelscript.psi.*;
 import com.aesirinteractive.angelscript.AngelscriptPsiImplUtil;
+import com.intellij.psi.PsiReference;
 
-public class AngelscriptVariableAccessExprImpl extends AngelscriptExprImpl implements AngelscriptVariableAccessExpr {
+public class AngelscriptVariableAccessExprImpl extends AngelscriptVariableAccessExprMixin implements AngelscriptVariableAccessExpr {
 
   public AngelscriptVariableAccessExprImpl(@NotNull ASTNode node) {
     super(node);
   }
 
-  @Override
   public void accept(@NotNull AngelscriptVisitor visitor) {
     visitor.visitVariableAccessExpr(this);
   }