+++ src/main/kotlin/com/aesirinteractive/angelscript/AngelscriptLsp.kt
@@ -12,69 +12,12 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
-import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.actionSystem.DataContext
-import com.intellij.openapi.editor.Editor
-import com.intellij.platform.lsp.api.LspServer
 import com.intellij.platform.lsp.api.LspServerManager
 import com.intellij.platform.lsp.api.LspServerSupportProvider
 import com.intellij.platform.lsp.api.ProjectWideLspServerDescriptor
 import com.intellij.platform.lsp.api.customization.LspCustomization
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
-import com.intellij.refactoring.actions.RenameElementAction;
-import com.intellij.refactoring.rename.RenameHandler
 
 private val LOG = logger<AngelscriptLspServerSupportProvider>()
-
-class AngelscriptRenameHandler : RenameHandler {
-    override fun isAvailableOnDataContext(dataContext: DataContext): Boolean {
-        val file = CommonDataKeys.PSI_FILE.getData(dataContext) ?: return false
-//        return file.language == AngelscriptLanguage.INSTANCE
-        return true
-    }
-
-    override fun invoke(
-        project: Project,
-        editor: Editor?,
-        file: PsiFile?,
-        dataContext: DataContext?
-    ) {
-        if (editor == null) {
-            return
-        }
-        val position = editor.caretModel.primaryCaret.logicalPosition
-        val ServerManager = LspServerManager.getInstance(project)
-        val AngelscriptLs: Collection<LspServer> = ServerManager.getServersForProvider(AngelscriptLspServerSupportProvider::class.java)
-//        AngelscriptLs.stream().findFirst().ifPresent { lspServer: LspServer ->
-//            lspServer.sendRequestSync { server: Lsp4jServer ->
-//                val fut = server.textDocumentService.rename(RenameParams(
-//                    TextDocumentIdentifier("test.as"),
-//                    Position(position.line, position.column),
-//                    "newName"
-//                ))
-//                fut
-//            }
-
-//        }
-        NotificationGroupManager.getInstance()
-            .getNotificationGroup("AngelScript LSP")
-            .createNotification("Custom rename 1", "Custom rename", NotificationType.INFORMATION)
-            .notify(project)
-    }
-
-    override fun invoke(
-        project: Project,
-        editor: Array<out PsiElement?>,
-        dataContext: DataContext?
-    ) {
-        NotificationGroupManager.getInstance()
-            .getNotificationGroup("AngelScript LSP")
-            .createNotification("Custom rename 2", "Custom rename", NotificationType.INFORMATION)
-            .notify(project)
-    }
-
-}
 
 internal class AngelscriptLspServerSupportProvider : LspServerSupportProvider {
     private class AngelLspServerDescriptor(project: Project) : ProjectWideLspServerDescriptor(project, "AngelScript") {