+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - Parse and highlight interfaces
 - Add option to use specific `.clang-format` file for AngelScript
 - Focus Rider when hitting breakpoints
+- Add support for file structure view (showing all symbols declared in the current file)
+- Add support for renamed local variables and types declared in AngelScript.